Add booting K2_CL (Boost)
K2_CL is now booting. Source based off of the k2_ul. Bugs should be reported ass issues on my github. Doesnt work on ICS (known issue) because the msm_fb drivers are not compatable. This is for JB and above only. Issues reported with ICS will be ignored.

This is fully compatable with the JB Sense rom (cricket) ported to the Boost variant.